pyproject.toml
[project]
name = "fuzzthat"
version = "0.1.0"
description = "State-Aware Smart Contract Fuzzing for EVM-based Projects."
authors = [
  { name="Maulana Malik Ibrahim", email="ceo@ibrahimstudio.com" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "web3",
  "hypothesis",
  "py-solc-x",
  "typer[all]",
  "pytest"
]

[project.scripts]
fuzzthat = "cli:app"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
